Move identity and new for matrices to `default.rs`
Also renamed `test_set_one` -> `test_identity`, otherwise the code is not modified.
