+++ src/integer/mat_poly_over_z.rs
@@ -14,6 +14,7 @@ use flint_sys::fmpz_poly_mat::fmpz_poly_mat_struct;
 mod arithmetic;
 mod cmp;
 mod concat;
+mod default;
 mod evaluate;
 mod from;
 mod get;
