+++ src/integer/mat_z/from.rs
@@ -19,65 +19,12 @@ use crate::{
     integer::Z,
     integer_mod_q::MatZq,
     traits::{GetNumColumns, GetNumRows, SetEntry},
-    utils::{
-        dimensions::find_matrix_dimensions, index::evaluate_index, parse::parse_matrix_string,
-    },
+    utils::{dimensions::find_matrix_dimensions, parse::parse_matrix_string},
 };
-use flint_sys::fmpz_mat::{
-    fmpz_mat_one, {fmpz_mat_init, fmpz_mat_set},
-};
-use std::{fmt::Display, mem::MaybeUninit, str::FromStr};
+use flint_sys::fmpz_mat::fmpz_mat_set;
+use std::str::FromStr;
 
 impl MatZ {
-    /// Creates a new matrix with `num_rows` rows, `num_cols` columns and
-    /// zeros as entries.
-    ///
-    /// Parameters:
-    /// - `num_rows`: number of rows the new matrix should have
-    /// - `num_cols`: number of columns the new matrix should have
-    ///
-    /// Returns a [`MatZ`] or an error, if the number of rows or columns is
-    /// less or equal to `0`.
-    ///
-    /// # Examples
-    /// ```
-    /// use qfall_math::integer::MatZ;
-    ///
-    /// let matrix = MatZ::new(5, 10).unwrap();
-    /// ```
-    ///
-    /// # Errors and Failures
-    /// - Returns a [`MathError`] of type
-    /// [`InvalidMatrix`](MathError::InvalidMatrix)
-    /// if the number of rows or columns is `0`.
-    /// - Returns a [`MathError`] of type [`OutOfBounds`](MathError::OutOfBounds)
-    /// if the number of rows or columns is negative or it does not fit into an [`i64`].
-    pub fn new(
-        num_rows: impl TryInto<i64> + Display,
-        num_cols: impl TryInto<i64> + Display,
-    ) -> Result<Self, MathError> {
-        let num_rows_i64 = evaluate_index(num_rows)?;
-        let num_cols_i64 = evaluate_index(num_cols)?;
-
-        if num_rows_i64 == 0 || num_cols_i64 == 0 {
-            return Err(MathError::InvalidMatrix(
-                "A matrix can not contain 0 rows or 0 columns".to_string(),
-            ));
-        }
-
-        // initialize variable with MaybeUn-initialized value to check
-        // correctness of initialization later
-        let mut matrix = MaybeUninit::uninit();
-        unsafe {
-            fmpz_mat_init(matrix.as_mut_ptr(), num_rows_i64, num_cols_i64);
-
-            // Construct MatZ from previously initialized fmpz_mat
-            Ok(MatZ {
-                matrix: matrix.assume_init(),
-            })
-        }
-    }
-
     /// Create a [`MatZ`] from a [`MatZq`].
     ///
     /// Parameters:
@@ -99,37 +46,6 @@ impl MatZ {
         let mut out = MatZ::new(matrix.get_num_rows(), matrix.get_num_columns()).unwrap();
         unsafe { fmpz_mat_set(&mut out.matrix, &matrix.matrix.mat[0]) };
         out
-    }
-
-    /// Generate a `num_rows` times `num_columns` matrix with `1` on the
-    /// diagonal and `0` anywhere else.
-    ///
-    /// Parameters:
-    /// - `rum_rows`: the number of rows of the identity matrix
-    /// - `num_columns`: the number of columns of the identity matrix
-    ///
-    /// Returns a matrix with `1` across the diagonal and `0` anywhere else.
-    ///
-    /// # Examples
-    /// ```
-    /// use qfall_math::integer::MatZ;
-    ///
-    /// let matrix = MatZ::identity(2, 3).unwrap();
-    ///
-    /// let identity = MatZ::identity(10, 10).unwrap();
-    /// ```
-    ///
-    /// # Errors and Failures
-    /// - Returns a [`MathError`] of type [`InvalidMatrix`](MathError::InvalidMatrix) or
-    /// [`OutOfBounds`](MathError::OutOfBounds) if the provided number of rows and columns
-    /// are not suited to create a matrix. For further information see [`MatZ::new`].
-    pub fn identity(
-        num_rows: impl TryInto<i64> + Display,
-        num_cols: impl TryInto<i64> + Display,
-    ) -> Result<Self, MathError> {
-        let mut out = MatZ::new(num_rows, num_cols)?;
-        unsafe { fmpz_mat_one(&mut out.matrix) };
-        Ok(out)
     }
 }
 
@@ -192,42 +108,6 @@ impl From<&MatZq> for MatZ {
 }
 
 #[cfg(test)]
-mod test_new {
-    use crate::{
-        integer::{MatZ, Z},
-        traits::GetEntry,
-    };
-
-    /// Ensure that entries of a new matrix are `0`.
-    #[test]
-    fn entry_zero() {
-        let matrix = MatZ::new(2, 2).unwrap();
-
-        let entry1 = matrix.get_entry(0, 0).unwrap();
-        let entry2 = matrix.get_entry(0, 1).unwrap();
-        let entry3 = matrix.get_entry(1, 0).unwrap();
-        let entry4 = matrix.get_entry(1, 1).unwrap();
-
-        assert_eq!(Z::ZERO, entry1);
-        assert_eq!(Z::ZERO, entry2);
-        assert_eq!(Z::ZERO, entry3);
-        assert_eq!(Z::ZERO, entry4);
-    }
-
-    /// Ensure that a new zero matrix fails with `0` as input.
-    #[test]
-    fn error_zero() {
-        let matrix1 = MatZ::new(1, 0);
-        let matrix2 = MatZ::new(0, 1);
-        let matrix3 = MatZ::new(0, 0);
-
-        assert!(matrix1.is_err());
-        assert!(matrix2.is_err());
-        assert!(matrix3.is_err());
-    }
-}
-
-#[cfg(test)]
 mod test_from_mat_zq {
     use crate::{
         integer::{MatZ, Z},
@@ -263,58 +143,6 @@ mod test_from_mat_zq {
         assert_eq!(Z::from(u64::MAX - 58), matz_1.get_entry(0, 0).unwrap());
         assert_eq!(Z::from(u64::MAX - 1), matz_2.get_entry(0, 1).unwrap());
         assert_eq!(Z::from(u64::MAX - 58), matz_2.get_entry(0, 0).unwrap());
-    }
-}
-
-#[cfg(test)]
-mod test_set_one {
-    use crate::{
-        integer::{MatZ, Z},
-        traits::GetEntry,
-    };
-
-    /// Tests if an identity matrix is set from a zero matrix.
-    #[test]
-    fn identity() {
-        let matrix = MatZ::identity(10, 10).unwrap();
-
-        for i in 0..10 {
-            for j in 0..10 {
-                if i != j {
-                    assert_eq!(Z::ZERO, matrix.get_entry(i, j).unwrap());
-                } else {
-                    assert_eq!(Z::ONE, matrix.get_entry(i, j).unwrap())
-                }
-            }
-        }
-    }
-
-    /// Tests if function works for a non-square matrix
-    #[test]
-    fn non_square_works() {
-        let matrix = MatZ::identity(10, 7).unwrap();
-
-        for i in 0..10 {
-            for j in 0..7 {
-                if i != j {
-                    assert_eq!(Z::ZERO, matrix.get_entry(i, j).unwrap());
-                } else {
-                    assert_eq!(Z::ONE, matrix.get_entry(i, j).unwrap())
-                }
-            }
-        }
-
-        let matrix = MatZ::identity(7, 10).unwrap();
-
-        for i in 0..7 {
-            for j in 0..10 {
-                if i != j {
-                    assert_eq!(Z::ZERO, matrix.get_entry(i, j).unwrap());
-                } else {
-                    assert_eq!(Z::ONE, matrix.get_entry(i, j).unwrap())
-                }
-            }
-        }
     }
 }
