+++ src/integer_mod_q/mat_zq.rs
@@ -15,6 +15,7 @@ use flint_sys::fmpz_mod_mat::fmpz_mod_mat_struct;
 mod arithmetic;
 mod cmp;
 mod concat;
+mod default;
 mod from;
 mod get;
 mod ownership;
